add istio 1.5 support

+++ otddctl.sh
@@ -2,16 +2,16 @@
 
 usage(){
 	echo "Usage: otddctl.sh [apply|delete] options"
-	echo -e "\t-v istioVersion: \t\tthe istio version when using apply command"
 	echo -e "\t-t targetDeployment: \t\tthe target deployment to record"
 	echo -e "\t-p port: \t\t\tthe deployment's container port"  
 	echo -e "\t-n namespace: \t\t\tthe target depoyment's namespace. default value: default"
 	echo -e "\t-i interval: \t\t\tthe redirector's interval(in milisecond) to redirect the request. defaults value: 1000"
 	echo -e "\t-P protocol: \t\t\tthe target depoyment's protocol. defaults value: http"
+        echo -e "\t-v istioVersion (optional): \t\tspecify the istio version"
 	echo ""
-	echo -e "\texamples: \t\t\totddctl.sh apply -v 1.2.2 -t reviews-v2 -p 9080 -n default -i 1000 -P http"
+	echo -e "\texamples: \t\t\totddctl.sh apply -t reviews-v2 -p 9080 -n default -i 1000 -P http"
 	echo -e "\t\t\t\t\totddctl.sh delete -t reviews-v2 -p 9080 -n default -i 1000 -P http"
-	echo -e "\t\t\t\t\totddctl.sh apply -v 1.2.2 -t reviews-v2 -p 9080"
+	echo -e "\t\t\t\t\totddctl.sh apply -t reviews-v2 -p 9080 -v 1.2.2"
 	echo -e "\t\t\t\t\totddctl.sh delete -t reviews-v2 -p 9080"
 	exit 1
 }
@@ -48,7 +48,7 @@ while getopts v:t:p:n:i:P: option
 do 
  case "${option}" 
  in 
- v) ISTIO_VERSION=${OPTARG};; 
+ v) SPECIFIED_ISTIO_VERSION=${OPTARG};;
  t) TARGETDEPLOYMENT=${OPTARG};; 
  p) PORT=${OPTARG};; 
  n) NAMESPACE=${OPTARG};; 
@@ -56,12 +56,6 @@ do
  P) PROTOCOL=${OPTARG};; 
  esac 
 done
-
-
-if [ -z "$ISTIO_VERSION" ] && [ $ACTION != "delete" ]
-then
-  usage
-fi
 
 if [ -z "$TARGETDEPLOYMENT" ]
 then
@@ -121,25 +115,69 @@ template=`echo "$template" | sed "s/{{NAMESPACE}}/$NAMESPACE/g"`
 template=`echo "$template" | sed "s/{{INTERVAL}}/$INTERVAL/g"`
 template=`echo "$template" | sed "s/{{PROTOCOL}}/$PROTOCOL/g"`
 
+if [ $ACTION == "apply" ]
+then
+
+  echo "determing installed istio version ..."
+  ISTIO_VERSION=`kubectl -n istio-system get configmap istio-sidecar-injector -o jsonpath='{.data.values}'|jq ".global.tag"`
+  #trim prefix/suffix quote
+  ISTIO_VERSION="${ISTIO_VERSION%\"}"
+  ISTIO_VERSION="${ISTIO_VERSION#\"}"
+
+  echo "installed istio version is $ISTIO_VERSION"
+
+  DIGITS=$(echo $ISTIO_VERSION | tr "\." "\n")
+  INDEX=0
+  for DIGIT in $DIGITS
+  do
+          if [ $INDEX -eq 0 ]; then
+                  MAJOR_VERSION=$DIGIT
+          fi
+          if [ $INDEX -eq 1 ]; then
+                  MINOR_VERSION=$DIGIT
+          fi
+          INDEX=$(( $INDEX + 1 ))
+  done
+
+  #before 1.5.0, it's an standalone istio-sidecar-injector to inject the istio proxy.  
+  if [[ $MAJOR_VERSION -eq 1 && $MINOR_VERSION -lt 5  ]]; then 
+    STIO_INSTANCE="istio-sidecar-injector"
+  else
+    ISTIO_INSTANCE="istiod"
+  fi
+
+if [[ ! -z "$SPECIFIED_ISTIO_VERSION " ]]
+then
+  echo "applying recorder for specific istio version $SPECIFIED_ISTIO_VERSION"
+  ISTIO_VERSION=$SPECIFIED_ISTIO_VERSION
+fi
+
+if [[ ! $ISTIO_VERSION =~ ^[0-9]+\.[0-9]+\.[0-9]+$ ]]; then
+  echo "version $ISTIO_VERSION is not supported! only three digit type of version is supported, e.g. 1.4.2 or 1.5.0"
+  exit 1
+fi
+
+fi
 
 if [ $ACTION == "apply" ]
 then
+
   #replace istio proxyv2 image to otdd proxyv2 to include otdd redirect/recorder plugins by changing the .Values.global.proxy.image in the configmap.
   #write to a temp file then VALUE=`cat .otdd_tmp instead of using VALUE=`...` directly because the \n will lost wiredly.
   kubectl -n istio-system get configmap istio-sidecar-injector -o json | jq ".data.config" |awk -v version="$ISTIO_VERSION" '{gsub(/name:[ ]+istio-proxy.+[ ]{4}ports/,"name: istio-proxy\\n    image: docker.io/otdd/proxyv2:" version "-otdd.0.1.0\\n    ports"); print $0}' > .otdd_tmp
   VALUE=`cat .otdd_tmp`
   rm -rf .otdd_tmp
-  INJECTOR_BEFORE=`kubectl -n istio-system get pods |grep istio-sidecar-injector|awk '{print $1}'`
+  INJECTOR_BEFORE=`kubectl -n istio-system get pods |grep $ISTIO_INSTANCE|awk '{print $1}'`
   echo "changing istio-sidecar-injector config and restart it.."
   kubectl -n istio-system get configmap istio-sidecar-injector -o json | jq ".data.config=$VALUE" | kubectl apply -f -
-  kubectl -n istio-system rollout restart deploy/istio-sidecar-injector
+  kubectl -n istio-system rollout restart deploy/$ISTIO_INSTANCE
   FOUND="1"
   while [ $FOUND == "1" ]
   do
     FOUND="0"
-    echo "waiting istio-sidecar-injector to be fully ready.. please do not kill this script."
+    echo "waiting $ISTIO_INSTANCE to be fully ready.. please do not kill this script."
     sleep 3
-    INJECTOR_AFTER=`kubectl -n istio-system get pods |grep istio-sidecar-injector|awk '{print $1}'`
+    INJECTOR_AFTER=`kubectl -n istio-system get pods |grep $ISTIO_INSTANCE|awk '{print $1}'`
     for i in $INJECTOR_BEFORE ; do
       for j in $INJECTOR_AFTER ; do
         if [[ $i == $j ]]
@@ -162,17 +200,17 @@ then
   kubectl -n istio-system get configmap istio-sidecar-injector -o json | jq ".data.config" |awk '{gsub(/name:[ ]+istio-proxy.+[ ]{4}ports/,"name: istio-proxy\\n  {{- if contains \\\"/\\\" .Values.global.proxy.image }}\\n    image: \\\"{{ annotation .ObjectMeta `sidecar.istio.io/proxyImage` .Values.global.proxy.image }}\\\"\\n  {{- else }}\\n    image: \\\"{{ annotation .ObjectMeta `sidecar.istio.io/proxyImage` .Values.global.hub }}/{{ .Values.global.proxy.image }}:{{ .Values.global.tag }}\\\"\\n  {{- end }}\\n    ports"); print $0}' > .otdd_tmp_uninstall
   VALUE=`cat .otdd_tmp_uninstall`
   rm -rf .otdd_tmp_uninstall
-  INJECTOR_BEFORE=`kubectl -n istio-system get pods |grep istio-sidecar-injector|awk '{print $1}'`
+  INJECTOR_BEFORE=`kubectl -n istio-system get pods |grep $ISTIO_INSTANCE|awk '{print $1}'`
   echo "restoring back istio-sidecar-injector config and restart it.."
   kubectl -n istio-system get configmap istio-sidecar-injector -o json | jq ".data.config=$VALUE" | kubectl apply -f -
-  kubectl -n istio-system rollout restart deploy/istio-sidecar-injector
+  kubectl -n istio-system rollout restart deploy/$ISTIO_INSTANCE
   FOUND="1"
   while [ $FOUND == "1" ]
   do
     FOUND="0"
-    echo "waiting istio-sidecar-injector to be fully ready.. "
+    echo "waiting $ISTIO_INSTANCE to be fully ready.. "
     sleep 3
-    INJECTOR_AFTER=`kubectl -n istio-system get pods |grep istio-sidecar-injector|awk '{print $1}'`
+    INJECTOR_AFTER=`kubectl -n istio-system get pods |grep $ISTIO_INSTANCE|awk '{print $1}'`
     for i in $INJECTOR_BEFORE ; do
       for j in $INJECTOR_AFTER ; do
         if [[ $i == $j ]]
